backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-google-genai
google-generativeai
google-api-python-client
google-auth
python-dotenv
